ImGui: Convert project to Alire
Converting to an Alire published crate will make this project
easier to build and publish. Since it only depends on 1 already
existing published alire crate, it seem logical to pull that
dependency from alire, and publish this to it to make it more widely
available.

Currently the libraries build, but the example project is still a WIP.